Icon: follow create component best practices.
It is considered unsafe to render a component directly from the
class or function if you're getting it from props.
The best practice for that is to create the element with
React and pass it around.
This fix will should remove many warnings about rendering custom
components that are not following the PascalCase naming.