feat(deps): Upgrade testing-library and user-event
Upgrades @testing-library/react to v13.4.0, which is the latest version supporting React 16/17.
Removes @testing-library/react-hooks as it is now included in @testing-library/react.
Aligns @testing-library/user-event to v13.5.0.
Aligns @testing-library/jest-dom to v6.4.6.

Also downgrades several @typescript-eslint packages and eslint-plugin-vitest to resolve an installation issue with the node version used in the project.